update/fix restart minikute issue - clean up residues files by Minikube

+++ run-Hello-Minikube.sh
@@ -1,6 +1,6 @@
 #!/bin/bash -x
 
-#minikube start
+minikube start
 
 kubectl run hello-minikube --image=k8s.gcr.io/echoserver:1.4 --port=8080
 
@@ -14,11 +14,59 @@ kubectl get pod
 # We can see that the pod is still being created from the ContainerCreating status
 kubectl get pod
 
+echo "..... Wait for 60 seconds for Kubernettes to come up fully ....."
+wait 60
+
 # We can see that the pod is now Running and we will now be able to curl it:
 curl $(minikube service hello-minikube --url)
 
-kubectl delete service hello-minikube
+echo "--------------------------------------------"
+echo "##### Way-4: Ask Yes/No with default Y #####"
+echo "--------------------------------------------"
+## arg1: "Prompt String, e.g. Enter choice (Y/n) default=Y? "
+## arg2: Default value, Y (default) or N
+answer=y
+function askYesNo() {
+    DEFAULT_YES_NO="y"
+    if [ $# -lt 2 ]; then
+        echo "---- INFO ----: No default choice provided, use Yes/Y as default!"
+    fi
+    # Prompt and wait for 5 seconds as timeout to accept default value.
+    #read -t 5 -n 1 -p "Enter choice (Y/n) default=Y? " answer
+    defaultValue=${2:-${DEFAULT_YES_NO}}
+    defaultValue=`echo ${defaultValue:0:1}| tr '[:upper:]' '[:lower:]'`
+    if [[ ! "nNyY" =~ "${defaultValue}" ]]; then
+        defaultValue="${DEFAULT_YES_NO}"
+    fi
+    #read -t 5 -n 1 -p "$1" answer
+    read -n 1 -p "$1" answer
+    [ -z "$answer" ] && answer="$defaultValue" 
+    answer=`echo ${answer:0:1}| tr '[:upper:]' '[:lower:]'`
+    case ${answer:0:1} in
+        y|Y )
+            echo "===> Yes (you choice $answer)"
+        ;;
+        n|N )
+            echo "===> No (you choice $answer)"
+        ;;
+        * ) 
+            echo "===> Yes (you choice $defaultValue)"
+            answer=$defaultValue
+        ;;
+    esac
+}
 
-kubectl delete deployment hello-minikube
+askYesNo "Continue to DELETE hello-minikube (N/y) default=No/N? " "N"
+if [ "$answer" == "y" ]; then
+    kubectl delete service hello-minikube
+    kubectl delete deployment hello-minikube
+fi
 
-minikube stop
+askYesNo "Continue to STOP -minikube (N/y) default=No/N? " "N"
+if [ "$answer" == "y" ]; then
+    minikube stop
+fi
+
+
+
+
